[ADD] added basic rendering to directionalityIndicator algorithm. Hardcoded mostly. Will be replaced by a proper OpenGL abstraction next.

+++ src/core/data/TriangleMesh.cpp
@@ -44,7 +44,7 @@ namespace di
             return m_triangles.size() - 1;
         }
 
-        size_t TriangleMesh::addVertex( double x, double y, double z )
+        size_t TriangleMesh::addVertex( float x, float y, float z )
         {
             return addVertex(
                 {
